region directives are now also moved accordingly by MiKo_4004
(resolves #725)
